pyproject.toml
[project]
name = "bookly"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.17.1",
    "asyncpg>=0.30.0",
    "bcrypt==4.0.1",
    "fastapi-mail>=1.5.8",
    "fastapi[standard]>=0.120.2",
    "itsdangerous>=2.2.0",
    "passlib>=1.7.4",
    "pydantic-settings>=2.11.0",
    "python-jose>=3.5.0",
    "redis>=7.0.1",
    "sqlalchemy[asyncio]>=2.0.44",
]
